// Frontend/types/LaptopFormData.interface.ts
export interface LaptopSpecs {
    comp: string; // Brand of the laptop
    type: string; // Type of usage, e.g., Gaming, Work, etc.
    inch: number; // Screen size in inches
    res: string; // Screen resolution 
    cpu: string; // Processor model
    gpu: string; // Graphics card brand/model

    os: string; // Operating system
    weight: number; // Weight in kg
  }